fsc/tokenizer: the token types are the grammar's; nothing from js/tokenizer
Stage 5c of todo/parser-serializer-restructure.md. The front end's
tokenizer was already the LL(1) grammar fjs/ebnf/lib/js plus a fold; what
it still took from the hand-written fjs/js/tokenizer was two helpers and
the token types. The keyword lookup the fold already had from
fjs/js/keywords, and the trivia rule is two lines the fold now states in
place; isKeywordToken, consumerless, is removed from js/tokenizer.

JsToken and its members, TokenMetadata, TokenPosition, TriviaKind,
JsTokenWithMetadata and the error-message union move to
fjs/ebnf/lib/js/types.ts, the grammar's vocabulary, with the operator kinds
derived from the grammar's own operators list rather than spelled again,
as a union of one-kind members so a switch on kind still narrows.
fjs/js/tokenizer/types.ts keeps the scanner's state types and reads the
vocabulary from the grammar; fjs/media/json/tokenizer reads it from there
too. No behavior changes.
